--- demo/form_graph/requirements.txt
@@ -1,1 +1,0 @@
-matplotlib
